Moved Motor to its own class
